issue with inserting sample tweets
