+++ .cfignore
@@ -1,5 +1,7 @@
 node_modules
 *.md
-dist
-.env
-src/test/*
+./server/dist
+./server/.env
+./server/test
+./client/node_modules
+./server/node_modules
