+++ server/src/routes/TweeterRoute.ts
@@ -2,16 +2,17 @@ import * as express from 'express';
 import { Request, Response } from 'express';
 import { TwitterOptions } from '../model/CRMModel';
 import { TweeterListener } from '../service/TweeterListener';
+import { EnrichmentPipeline } from '../util/EnrichmentPipeline';
 
 export class TweeterRoute {
 
   public router: express.Router = express.Router();
   private tweeterListener: TweeterListener;
 
-  constructor() {
+  constructor(enrichmentPipeline: EnrichmentPipeline) {
     this.routes();
     const twitOptions: TwitterOptions = {} as TwitterOptions;
-    this.tweeterListener = TweeterListener.getInstance(twitOptions);
+    this.tweeterListener = TweeterListener.getInstance(twitOptions,  enrichmentPipeline);
   }
 
   private routes(): void {
